Updated the TODO lists.
Fully implemented user, mod and admin commands. The bot is now able to
react based on text commands. Feature sets are implemented into the bot.

Split the config file up into the connection and channel components.
Configuration settings for connection never change and are static. Channel
config files will contain the dynamic, per-channel settings and will be
generated when a new instance of pyayaBot is spun up and added to a
channel. Revamped the Userclass to allow checking, getting and setting
admin, broadcaster and op status. I still need to go through to make sure
I've replaced code in main and featureSets with method calls. The mod list
is now updated in real-time. Part of the process of updating the list is
to also add or remove the op status from the corresponding User object.
The admin list will be able to contain broadcaster-specified usernames.
The same CRUD operations are available for the admin list as in the op
list. Added more logging and comments. Added print methods for most
classes. Green comment print method calls need to be replaced by DEBUG-level system logging.

I think that's everything for now. :)